refactor: change backend folder structure
